Change Default behavior to return observations
The default post will not return a JSON object that contains
observations, as opposed to the previous list style. As of now only
findPredforPredatorObseravation is implemented. Meaning that if the user
wants all of the prey observations of a particular predator it will
work. The inverse of that will not. You cannot filter by a specific prey
type yet. Also added all the test for this. Both Mock and REST should be
working